Created concept of node manager which is a console side proxy class to enable translation of commands between UI and mixnet. This layer can be used to enforce defaults and to isolate the mixnet from the UI.
Console uses reflection to discover commands which are annotated with ConsoleCommand and ConsoleParam to define command parameters.
With this adding new commands does not require any special changes to the UI the ui will simply pick up on the details.

There is an example in MixnetCommandServicesAdapter.

Javascript on the UI side will render the command dialog.

The console project has a Main class, which opens a jetty instance on Port 1887 bound to 0.0.0.0

There is not real configuration as yet and any commands will return filler.

The command dialog on the UI side are not rendered yet.

Megan