Gutted installer have basic recursive directory copying.
There is code left in place to ask questions and there is an interface and classes that can be extended to add new user inputs.
The abstraction exists so that the same tasks can be used in the UI version or the console version.

If a command line argument is provided the installer will use that to find the jar, which is useful for testing otherwise it locates
itself.

For this project the Gradle jar task has been exteded, within that extension developers may add more files that will be included in the installer.

Commented out in the prohect main gradle.build file is a task "make_installer" that will is partway built but enables administators to build
installers and control what the user is prompted to submit.

The core of this installer is a framework for an interactive and configurable installer.

No attempt at signing or validation has been implemented.

The UI interaction is complted, the console interaction is a stub.

Running the installer at present gives the user no feedback.

Take a look at "installer/resources/install.xml"

Cheers